chore: return empty buffer for files with no data
Part of making the hases emitted by js-ipfs the same as go-ipfs meant unsetting
the data attribute of UnixFS metadata objects if the data buffer had 0 length.

ipfs.files.get converts a pull stream of files into a readable stream of files
and all of the content pull streams too. If an empty buffer is not emitted by
the file content pull stream for file with no data then no file is written
out to the filesystem.

This was a fun bug to track down.

This module used to emit empty buffers for UnixFS metadata with no data
property but I removed it as part of #208 as there was no test that failed
without it so here we reinstate it and add a test.