fix: return cids from builder
Enforces CID Version 1 for non-sha2-256 hash algorithms and returns
the buffer that makes up the CID instead of the buffer that makes
up the multihash.

Returning the multihash for v0 CIDs is fine because the codec and
version are implicity but if we want to be able to load DAGNodes
stored with a v1 CID, we need to return the CID buffer instead.