feat: add `fullPath` option
When exporting a path with a subtree it's sometimes useful to get
information about everything in the path and not just the node
at the end.

This PR adds an option to return intermediate nodes in the export
stream.

It also documents the `maxDepth` option and lowers the sharding
threshold to make the shard tests more reliable in the browser.