test: adds benchmark test
Because it's useful to have this somewhere. Skipped because we can use this to do adhoc testing and don't want it to run as part of CI.

It uses 'realistic' data sizes, so 64k incoming chunks and a 500MB file size.

It passes zeroed buffers as using crypto to generate random bytes was a bottleneck and didn't seem to add much to the test.

It runs the test 10 times then prints out the timing data collected while importing the file which you can then use to create graphs in Google Docs or whatever.