feat: adds metadata to unixfs
Adds two new fields `mtime` and `mode` in line with the unixfs spec.

One possible wart here is that `mtime` is implemented as a JavaScript
number which means it can't hold the full value of `int64` so cannot
represent dates in the far future.

This should be addressed by UnixFSv2 which should hopefully arrive
before we top out `Number.MAX_SAFE_INTEGER` - e.g. early June 2255.