feat: support aborting exports
Adds a `signal` option to the exporter that will be passed to `ipld.get`,
allowing it to be notified that the user is no longer interested in
exporting the file/directory/etc.

Current behaviour is that if the node is offline and a block is not in
the repo is will throw, if it's online the CID will be added to the
bitswap want list via the block service.

Follow up PRs to ipld, block service and bitswap will impelement the
removal logic.