fix: just do what the user asked (#90)
We change some defaults based on the CID version passed, this seems like the wrong approach and debugging problems turns into whack-a-mole.

Let the calling code decide the settings for the various knobs and levers instead.

BREAKING CHANGE: We previously did things like set rawLeaves to true based on the CID version, now we do not.