fix: types with ipjs build (#165)
Currently, the published package.json has `"types": "dist/src/index.d.ts"` and should be `"types": "src/index.d.ts",` according to where the types really are.

The problem here is that we need to both:
- have type paths correct in the dist folder
- have type paths correct for local dev in src

Given the path will change, we need to guarantee this is updated in the build `package.json`. Using an approach like `multiformats` and `uint8arrays` having a `types` folder which is moved to `dist` is not possible because of the inline `d.ts` files in this repo, which will be inside `dist` folder as `dist/src/*.d.ts`, breaking the imports from the types around it.

With the above in mind, this PR adds a postbuild script where the package.json is changed in post build to have `dist` removed.