fix: individual packages can use npm 6 (#167)
We have engines with npm7 because we use npm7 workspaces, but this should only be needed in the main package.json for development. Internal packages should be published as supporting npm6 given that there are linux environments that still do not support npm7. For more information see https://github.com/web3-storage/web3.storage/issues/387 where this issue was raised

Closes https://github.com/web3-storage/web3.storage/issues/387